+++ src/Collections/BacktraceCollection.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Stratedge\Wye\Collections;
+
+class BacktraceCollection extends Collection implements BacktraceCollectionInterface
+{
+
+}