+++ src/Collections/BacktraceCollectionInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Stratedge\Wye\Collections;
+
+interface BacktraceCollectionInterface extends CollectionInterface
+{
+
+}